修改Math.Round(526.925,2)为(decimal)526.925, 2,MidpointRounding.AwayFromZero)【四舍五入】

+++ 测试系统-数据管理模块/Models/ReadResult.cs
@@ -15,7 +15,7 @@ namespace 数据管理模块.Models
         //[SugarColumn(ColumnName = "测试项目")]
         public string  测试项目 {get;set; }
         //[SugarColumn(ColumnName = "测试结果")]
-        public double 测试结果 {get;set; }
+        public decimal 测试结果 {get;set; }
         //[SugarColumn(ColumnName = "UUT批次号",IsNullable =true)]
         public string UUT批次号 { get; set; }
         public bool IsPassed { get;set;}
